requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
schedule>=1.2.0
pytz>=2023.3
requests>=2.31.0
